refactor: hook -> util
