+++ src/lib/hooks/index.ts
@@ -5,13 +5,11 @@ export * from './useDomainData';
 export * from './useDomainEvents';
 export * from './useDomainMetadata';
 export * from './useDomainMetrics';
-export * from './useDownloadAsset';
 export * from './useIsDomainMetadataLocked';
 export * from './usePaymentToken';
 export * from './usePaymentTokenForDomain';
 export * from './usePaymentTokenInfo';
 export * from './useResize';
-export * from './useShareAsset';
 export * from './useSubdomainData';
 export * from './useUserTokenBalance';
 export * from './useViewNavigation';
